✅[CHORE]: 함수의 기능을 파일별로 분리 및 네이밍 수정
