+++ RockPaperScissors/RockPaperScissors/main.swift
@@ -6,4 +6,4 @@
 
 import Foundation
 
-printMenu()
+startProgram()
